+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ This document describes changes between each past release.
 - List of recipients can now contain groups URIs. The principals from the specified
   group that look like email addresses will be used as recipients (fixes #6)
 - Support new variables like server root url or client IP address in email template (fixes #22)
+- Add some validation when defining kinto-emailer settings in collections metadata (fixes #21)
 
 
 0.1.0 (2017-01-25)